--- inchworm_control/block_simulation/__init__.py
@@ -1,3 +1,0 @@
-from inchworm_control import blueprint as BP
-
-BP.blueprint()
