--- inchworm_control/inchworm_control/__init__.py
@@ -1,1 +1,0 @@
-from .blueprint import blueprint
