+++ inchworm_control/inchworm_project.egg-info/SOURCES.txt
@@ -3,6 +3,7 @@ setup.py
 block_simulation/__init__.py
 block_simulation/a_star_path_planning.py
 block_simulation/bfs_path_planning.py
+block_simulation/blueprint.py
 block_simulation/config.py
 block_simulation/convert.py
 block_simulation/inchworm_data.py
@@ -12,9 +13,7 @@ block_simulation/press_n.py
 block_simulation/search.py
 block_simulation/sim.py
 block_simulation/sim_data.py
-block_simulation/tracys_testing_corner.py
 inchworm_control/__init__.py
-inchworm_control/blueprint.py
 inchworm_control/ik.py
 inchworm_control/ik_test.py
 inchworm_control/iw_UART.py
