+++ inchworm_control/inchworm_project.egg-info/entry_points.txt
@@ -1,4 +1,5 @@
 [console_scripts]
+inchworm_data = block_simulation.inchworm_data:main
 motor_controller = inchworm_control.motor_controller:main
 step_publisher = inchworm_control.step_publisher:main
